+++ tests/e2e/playwright.config.ts
@@ -28,8 +28,9 @@ export default defineConfig({
   ],
 
   webServer: {
-    command: 'cd ../../ && npm run dev',
+    command: 'npm run dev',
     url: 'http://localhost:3000',
     reuseExistingServer: !process.env.CI,
+    cwd: '../..',
   },
 })